+++ pkger/models.go
@@ -275,28 +275,74 @@ func (d DiffCheck) IsNew() bool {
 	return d.Old == nil
 }
 
-// DiffDashboard is a diff of an individual dashboard. This resource is always new.
-type DiffDashboard struct {
+// DiffDashboardValues are values for a dashboard.
+type DiffDashboardValues struct {
 	Name   string      `json:"name"`
 	Desc   string      `json:"description"`
 	Charts []DiffChart `json:"charts"`
 }
 
+// DiffDashboard is a diff of an individual dashboard.
+type DiffDashboard struct {
+	ID      SafeID               `json:"id"`
+	Remove  bool                 `json:"remove"`
+	PkgName string               `json:"pkgName"`
+	New     DiffDashboardValues  `json:"new"`
+	Old     *DiffDashboardValues `json:"old"`
+}
+
 func newDiffDashboard(d *dashboard) DiffDashboard {
 	diff := DiffDashboard{
-		Name: d.Name(),
-		Desc: d.Description,
+		ID:      SafeID(d.ID()),
+		PkgName: d.PkgName(),
+		New: DiffDashboardValues{
+			Name:   d.Name(),
+			Desc:   d.Description,
+			Charts: make([]DiffChart, 0, len(d.Charts)),
+		},
 	}
 
 	for _, c := range d.Charts {
-		diff.Charts = append(diff.Charts, DiffChart{
+		diff.New.Charts = append(diff.New.Charts, DiffChart{
 			Properties: c.properties(),
 			Height:     c.Height,
 			Width:      c.Width,
 		})
 	}
 
+	if !d.Exists() {
+		return diff
+	}
+
+	oldDiff := DiffDashboardValues{
+		Name:   d.existing.Name,
+		Desc:   d.existing.Description,
+		Charts: make([]DiffChart, 0, len(d.existing.Cells)),
+	}
+
+	for _, c := range d.existing.Cells {
+		var props influxdb.ViewProperties
+		if c.View != nil {
+			props = c.View.Properties
+		}
+
+		oldDiff.Charts = append(oldDiff.Charts, DiffChart{
+			Properties: props,
+			XPosition:  int(c.X),
+			YPosition:  int(c.Y),
+			Height:     int(c.H),
+			Width:      int(c.W),
+		})
+	}
+
+	diff.Old = &oldDiff
+
 	return diff
+}
+
+// IsNew indicates whether the pkg dashboard is new to the platform.
+func (d DiffDashboard) IsNew() bool {
+	return d.Old != nil
 }
 
 // DiffChart is a diff of oa chart. Since all charts are new right now.
@@ -417,8 +463,7 @@ func (d DiffNotificationEndpoint) IsNew() bool {
 	return d.Old == nil
 }
 
-// DiffNotificationRule is a diff of an individual notification rule. This resource is always new.
-type DiffNotificationRule struct {
+type DiffNotificationRuleValues struct {
 	Name        string `json:"name"`
 	Description string `json:"description"`
 
@@ -430,29 +475,94 @@ type DiffNotificationRule struct {
 	Every           string              `json:"every"`
 	Offset          string              `json:"offset"`
 	MessageTemplate string              `json:"messageTemplate"`
-	Status          influxdb.Status     `json:"status"`
 	StatusRules     []SummaryStatusRule `json:"statusRules"`
 	TagRules        []SummaryTagRule    `json:"tagRules"`
 }
 
+// DiffNotificationRule is a diff of an individual notification rule. This resource is always new.
+type DiffNotificationRule struct {
+	ID      SafeID `json:"id"`
+	Remove  bool   `json:"bool"`
+	PkgName string `json:"pkgName"`
+
+	New DiffNotificationRuleValues  `json:"new"`
+	Old *DiffNotificationRuleValues `json:"old"`
+}
+
 func newDiffNotificationRule(r *notificationRule, iEndpoint influxdb.NotificationEndpoint) DiffNotificationRule {
 	sum := DiffNotificationRule{
-		Name:            r.Name(),
-		Description:     r.description,
-		EndpointName:    r.endpointName.String(),
-		Every:           r.every.String(),
-		Offset:          r.offset.String(),
-		MessageTemplate: r.msgTemplate,
-		Status:          r.Status(),
-		StatusRules:     toSummaryStatusRules(r.statusRules),
-		TagRules:        toSummaryTagRules(r.tagRules),
+		ID:      SafeID(r.ID()),
+		Remove:  r.shouldRemove,
+		PkgName: r.PkgName(),
+		New: DiffNotificationRuleValues{
+			Name:            r.Name(),
+			Description:     r.description,
+			EndpointName:    r.endpointName.String(),
+			Every:           r.every.String(),
+			Offset:          r.offset.String(),
+			MessageTemplate: r.msgTemplate,
+			StatusRules:     toSummaryStatusRules(r.statusRules),
+			TagRules:        toSummaryTagRules(r.tagRules),
+		},
 	}
 	if iEndpoint != nil {
-		sum.EndpointID = SafeID(iEndpoint.GetID())
-		sum.EndpointType = iEndpoint.Type()
+		sum.New.EndpointID = SafeID(iEndpoint.GetID())
+		sum.New.EndpointType = iEndpoint.Type()
+	}
+
+	if r.existing == nil {
+		return sum
+	}
+
+	sum.Old = &DiffNotificationRuleValues{
+		Name:         r.existing.rule.GetName(),
+		Description:  r.existing.rule.GetDescription(),
+		EndpointName: r.existing.endpointName,
+		EndpointID:   SafeID(r.existing.rule.GetEndpointID()),
+		EndpointType: r.existing.endpointType,
+	}
+
+	assignBase := func(b rule.Base) {
+		if b.Every != nil {
+			sum.Old.Every = b.Every.TimeDuration().String()
+		}
+		if b.Offset != nil {
+			sum.Old.Offset = b.Offset.TimeDuration().String()
+		}
+		for _, tr := range b.TagRules {
+			sum.Old.TagRules = append(sum.Old.TagRules, SummaryTagRule{
+				Key:      tr.Key,
+				Value:    tr.Value,
+				Operator: tr.Operator.String(),
+			})
+		}
+		for _, sr := range b.StatusRules {
+			sRule := SummaryStatusRule{CurrentLevel: sr.CurrentLevel.String()}
+			if sr.PreviousLevel != nil {
+				sRule.PreviousLevel = sr.PreviousLevel.String()
+			}
+			sum.Old.StatusRules = append(sum.Old.StatusRules, sRule)
+		}
+	}
+
+	switch p := r.existing.rule.(type) {
+	case *rule.HTTP:
+		assignBase(p.Base)
+	case *rule.Slack:
+		assignBase(p.Base)
+		sum.Old.MessageTemplate = p.MessageTemplate
+	case *rule.PagerDuty:
+		assignBase(p.Base)
+		sum.Old.MessageTemplate = p.MessageTemplate
 	}
 
 	return sum
+}
+
+// IsNew indicates if the resource will be new to the platform or if it edits
+// an existing resource.
+func (d DiffNotificationRule) IsNew() bool {
+	return d.Old != nil
 }
 
 // DiffTask is a diff of an individual task. This resource is always new.
@@ -1667,14 +1777,25 @@ type notificationRule struct {
 	endpointName *references
 	endpointType string
 
+	existing *existingRule
+
 	labels sortedLabels
 }
 
+type existingRule struct {
+	rule         influxdb.NotificationRule
+	endpointName string
+	endpointType string
+}
+
 func (r *notificationRule) Exists() bool {
-	return false
+	return r.existing != nil
 }
 
 func (r *notificationRule) ID() influxdb.ID {
+	if r.existing != nil {
+		return r.existing.rule.GetID()
+	}
 	return r.id
 }
 
@@ -2242,6 +2363,8 @@ type dashboard struct {
 	Charts      []chart
 
 	labels sortedLabels
+
+	existing *influxdb.Dashboard
 }
 
 func (d *dashboard) ID() influxdb.ID {
@@ -2257,7 +2380,7 @@ func (d *dashboard) ResourceType() influxdb.ResourceType {
 }
 
 func (d *dashboard) Exists() bool {
-	return false
+	return d.existing != nil
 }
 
 func (d *dashboard) summarize() SummaryDashboard {