+++ ui/src/dataLoaders/actions/dataLoaders.ts
@@ -621,12 +621,14 @@ export const writeLineProtocolAction = (
     } else if (resp.status === 429) {
       dispatch(notify(readWriteCardinalityLimitReached(resp.data.message)))
       dispatch(setLPStatus(RemoteDataState.Error))
+    } else if (resp.status === 403) {
+      dispatch(setLPStatus(RemoteDataState.Error, resp.data.message))
     } else {
+      dispatch(setLPStatus(RemoteDataState.Error, 'failed to write data'))
       throw new Error(get(resp, 'data.message', 'Failed to write data'))
     }
   } catch (error) {
     console.error(error)
-    dispatch(setLPStatus(RemoteDataState.Error, error.message))
   }
 }